Some state

+++ src/main/java/dk/htr/games/minmax/four_in_row/minimax/MiniMaxAlgorithm.java
@@ -4,5 +4,5 @@ import dk.htr.games.minmax.four_in_row.exceptions.GameException;
 
 public interface MiniMaxAlgorithm {
     int miniMax(long theBoard, boolean isMaximising) throws GameException;
-    int getNumberOfBoardsEvaluated();
+    long getNumberOfBoardsEvaluated();
 }
